Merge main and main_prod roles
The original idea to have a separate *_prod role for per-role production-only
tasks turns out not to make much sense. The original idea was to have database
dumps for production deployments and something else to create the database in
the dev env. However, we can still do db dumps for dev, except with "stub"
databases that don't contain any private data.