Define EXT variable locally
config.php has some autoloading magic at the end that loads stuff we
need. This can't be done in CI's autoload because our controllers
extend the stuff we need to load, and it's the base controller's
constructor that loads the Loader, which loads the stuff we extend. So
we need to autoload our classes before constructing our controllers.
This is what that magic bit of code does. It uses the EXT variable,
which used to be defined by old versions of CI but is no longer, so we
just define it ourselves locally.

Change-Id: If69b41ba9e912dc1de673fb61de672546822eddd