Support running dbscrub on any MySQL server
Previously, dbscrub assumed it was running on a LibriVox server
(normally it would be run on staging or localdev, for obvious
reasons). This patch adds support for specifying a username, password
and/or database to allow it to run on any MySQL server where the
un-scrubbed catalog has been imported.

Change-Id: I0b25bbdcc88203a1f7474fe54cbe8283f71900c1