Switch to `files` session driver
In CodeIgniter 2, sessions were stored in client-side encrypted
cookies [1]. This means we need to pick a session driver to store
sessions server-side. We initially went with memcache, but issue #7 is
reporting that it's not working too well. Swtich to the files driver
instead, and store the session files in the tmpfs-mounted /tmp

Resolves #7

[1] https://codeigniter.com/userguide3/installation/upgrade_300.html#step-6-update-your-session-library-usage