Don't use recurse for directories
`recurse` applies to the *contents*, not the path itself (in other
words, it's not the equivalent of the -p argument to `mkdir`). Ansible
will create the whole path by default. This was previously resetting
the permissions on the contents of the librivox-shared-validator-books
and uploads directories, which is not what we wanted.