Run unit tests in CI
Now that https://github.com/LibriVox/librivox-catalog/pull/190 has
merged, we can actually run in CI. We also lay down some groundwork to
make the workflow reusable, and to make the catalog branch checkout
configurable, with the intention of running it against unmerged PRs.
This means we have to call our own reusable workflow with the correct
value of catalog_git_branch set, which is weird, but I couldn't find
a more sane way to handle variable with GitHub Actions.